% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_indice_masa_corporal.R
\name{calcular_indice_masa_corporal}
\alias{calcular_indice_masa_corporal}
\title{Calculo del indice de masa corporal (IMC)}
\usage{
calcular_indice_masa_corporal(peso, altura)
}
\arguments{
\item{peso}{Peso en kilogramos. Debe ser numerico y mayor que cero.}

\item{altura}{Altura en metros. Debe ser numerica y mayor que cero.}
}
\value{
Un valor numerico correspondiente al IMC calculado.
}
\description{
Esta funcion calcula el IMC a partir del peso (en kilogramos)
y la altura (en metros). Valida que ambos valores sean numericos
y mayores que cero.
}
\examples{
calcular_indice_masa_corporal(3.5, 0.40)

}
