% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_pulgadas_a_centimetros.R
\name{convertir_pulgadas_a_centimetros}
\alias{convertir_pulgadas_a_centimetros}
\title{Convertir pulgadas a centímetros}
\usage{
convertir_pulgadas_a_centimetros(pulgadas)
}
\arguments{
\item{pulgadas}{Valor numérico que representa la longitud en pulgadas.}
}
\value{
Un número que indica la longitud convertida a centímetros.
}
\description{
Convierte una medida expresada en pulgadas a su equivalente en centímetros.
La función valida que el argumento de entrada sea numérico.
}
\examples{
convertir_pulgadas_a_centimetros(10)
convertir_pulgadas_a_centimetros(2.5)

}
