% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_indice_masa_corporal.R
\name{calcular_indice_masa_corporal}
\alias{calcular_indice_masa_corporal}
\title{Calcular el Índice de Masa Corporal (IMC)}
\usage{
calcular_indice_masa_corporal(peso, altura)
}
\arguments{
\item{peso}{Peso del individuo en kilogramos (numérico).}

\item{altura}{Altura del individuo en metros (numérico).}
}
\value{
Un número que representa el IMC calculado como \code{peso / altura^2}.
}
\description{
Calcula el Índice de Masa Corporal (IMC) a partir del peso y la altura.
La función verifica que ambos parámetros sean numéricos y mayores que cero.
}
\examples{
calcular_indice_masa_corporal(70, 1.75)
calcular_indice_masa_corporal(3.5, 0.40)

}
