% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_pulgadas_a_centimetros.R
\name{convertir_pulgadas_a_centimetros}
\alias{convertir_pulgadas_a_centimetros}
\title{Convertir pulgadas a centimetros}
\usage{
convertir_pulgadas_a_centimetros(pulgadas)
}
\arguments{
\item{pulgadas}{Valor numerico en pulgadas.}
}
\value{
Valor numerico convertido a centimetros.
}
\description{
Esta funcion convierte un valor numerico en pulgadas a centimetros.
}
\examples{
convertir_pulgadas_a_centimetros(10)

}
