% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_pulgadas_a_centimetros.R
\name{convertir_pulgadas_a_centimetros}
\alias{convertir_pulgadas_a_centimetros}
\title{Convertir pulgadas a centímetros}
\usage{
convertir_pulgadas_a_centimetros(pulgadas)
}
\arguments{
\item{pulgadas}{Un vector numérico con las medidas en pulgadas.}
}
\value{
Un vector numérico con las medidas convertidas a centímetros.
}
\description{
Convierte un vector de medidas de pulgadas a centímetros.
}
\examples{
convertir_pulgadas_a_centimetros(c(10, 15, 20))
}
